requirements.txt
annotated-types==0.7.0
anyio==4.5.2
bcrypt==4.3.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==45.0.5
decorator==5.2.1
Deprecated==1.2.18
exceptiongroup==1.3.0
fabric==3.2.2
fastapi==0.116.1
fpdf==1.7.2
h11==0.16.0
idna==3.10
invoke==2.2.0
Jinja2==3.1.6
logger==1.4
loguru==0.7.3
MarkupSafe==2.1.5
paramiko==3.5.1
pdfkit==1.0.0
pillow==10.4.0
pkg_resources==0.0.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyNaCl==1.5.0
python-multipart==0.0.20
python-nmap==0.7.1
reportlab==4.4.2
requests==2.32.4
sniffio==1.3.1
starlette==0.44.0
typing_extensions==4.13.2
urllib3==2.2.3
uvicorn==0.33.0
wrapt==1.17.2